requirements.txt
pysmb==1.1.18
docopt
pyyaml
